+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version)
 for commit guidelines.
 
+## 0.1.6 (2024-09-13)
+
+- Add support for deleting a service
+- Remove `httpx` logging handler
+- Add logger options for more personalized logging
+
 ## 0.1.5 (2024-09-02)
 
 - Execute a service using Transforms API